Create file_reading.cpp
I wanted to figure out how to count how many unique words appear in a document. Since I am familiar with the use of classes but not structs, I wanted to also figure out how to use structs. I will create a Node struct and try out a binary tree search/sort on the different words in the file. The Nodes will also track the frequency of word occurrence.